adc: add shift register clock/SCK offset
Without an offset, the output signal (SDI) will be set *as the ADC reads
it*, i.e. on the same clock edge. This results in the first bit being
read by the ADC as the second bit, and so on.

This is resolved by adding a 1/2-cycle offset (relative to the 25 MHz
communication clock) between the shift registers' clock and the one
transmitted to the ADC itself.